Create find_the_index
This came about from a tech test where I was asked to debug some code. What they were providing really seemed to go round the houses, so I thought I'd provide a bit of a replacement. Is it the most efficient solution? Probably not, but if I were them this is how I'd have done it. Does it show that I'm not stupid and could've come up with a solution? I hope so...